Make root disk size configurable
In the Docker container, the root disk is a cow_image.qcow2 image with
a size that previously was hard-coded to 50GB in the container start
script. Now that value can be configured (--disk parameter, "disk"
entry in the "size" struct of a VM instance when using a compose file).

When using a compose file that determines the size based on the
"flavor", the disk size defaults are increasing exponentially: 50GB
for "micro", 100GB for "tiny", etc.

Because the image file is a sparse file, this usually doesn't increase
actual disk usage unless the VM really writes much data.